refactor(app): split model.go (1103 lines) and update.go (821 lines)
model.go was a 1100-line catch-all on *model; update.go was 821. Both
have clear method clusters by responsibility, so split each into
descriptively-named sibling files. Pure code motion — no behaviour
changes.

model.go split:
  model_lifecycle.go     construction, run options, task wiring, shutdown
  model_session.go       session save/load, task storage, fork
  model_scrollback.go    render committed messages
  model_agent_events.go  conv.Runtime callbacks (turn lifecycle)
  model_compact.go       auto + /compact handling
  model_tool_effects.go  tool side effects (cwd, files, overflow)
  model_workspace.go     cwd / file change reactions + FileWatcher
  model_turn_queue.go    inbox drain + prompt injection at turn end
  model_deps.go          deps builders for sub-features
  model_actions.go       identity switch + slash command dispatch

update.go split:
  update_keys.go            keyboard handling + active-modal delegation
  update_resize.go          window resize + scrollback reflow
  update_submit.go          submit + provider turn + skill invocation
  update_command.go         slash command deps + execution
  update_modal.go           operation mode + question modal protocol
  update_approval.go        permission approval flow + bridge response
  update_input_effects.go   stream cancel, tool cancel, paste, quit

Each new file starts with a docstring describing what it owns; the
top of model.go / update.go lists all siblings as a navigation aid.
docs/packages/ui.md updated.
