+++ internal/app/model_deps.go
@@ -1,0 +1,73 @@
+// Deps builders: every sub-feature (input, conv, trigger) takes its
+// dependencies via a struct rather than reaching for package globals. These
+// methods assemble those structs from the model's services + env state.
+// Putting them in one place keeps the call sites in update.go / model.go
+// concise and makes the surface each sub-feature uses explicit.
+package app
+
+import (
+	tea "github.com/charmbracelet/bubbletea"
+
+	"github.com/genai-io/gen-code/internal/app/input"
+	"github.com/genai-io/gen-code/internal/app/kit"
+	"github.com/genai-io/gen-code/internal/app/trigger"
+	"github.com/genai-io/gen-code/internal/core"
+	"github.com/genai-io/gen-code/internal/llm"
+)
+
+func (m *model) promptSuggestionDeps() input.PromptSuggestionDeps {
+	return input.PromptSuggestionDeps{
+		Input:        &m.userInput,
+		Conversation: &m.conv.ConversationModel,
+		HasProvider:  m.env.LLMProvider != nil,
+		BuildClient:  m.buildLLMClient,
+	}
+}
+
+func (m *model) overlayDeps() input.OverlayDeps {
+	return input.OverlayDeps{
+		State:             &m.userInput,
+		Conv:              &m.conv.ConversationModel,
+		Cwd:               m.env.CWD,
+		CommitMessages:    m.CommitMessages,
+		CommitAllMessages: m.commitAllMessages,
+		SwitchProvider: func(p llm.Provider) {
+			m.StopAgentSession()
+			m.switchProvider(p)
+			m.ReconfigureAgentTool()
+		},
+		SetCurrentModel: func(info *llm.CurrentModelInfo) {
+			m.env.CurrentModel = info
+		},
+		ClearCachedInstructions: m.env.ClearCachedInstructions,
+		RefreshMemoryContext:    m.refreshMemoryContext,
+		FireFileChanged:         m.fireFileChanged,
+		ReloadPluginState:       m.ReloadPluginBackedState,
+		LoadSession:             m.loadSessionByID,
+		SetActiveIdentity:       m.setActiveIdentity,
+		DispatchSlashCommand:    m.dispatchSlashCommand,
+	}
+}
+
+func (m *model) triggerDeps() trigger.Deps {
+	return trigger.Deps{
+		StreamActive: m.conv.Stream.Active,
+		Cron:         m.services.Cron,
+		InjectCron:   m.injectCronPrompt,
+		InjectHook:   m.injectAsyncHookContinuation,
+		AppendNotice: func(text string) {
+			if text != "" {
+				m.conv.Append(core.ChatMessage{Role: core.RoleNotice, Content: text})
+			}
+		},
+	}
+}
+
+func (m *model) StartExternalEditor(path string) tea.Cmd {
+	return kit.StartExternalEditor(path, func(err error) tea.Msg {
+		return input.MemoryEditorFinishedMsg{Err: err}
+	})
+}
+
+func (m *model) SpinnerTickCmd() tea.Cmd { return m.conv.Spinner.Tick }
+func (m *model) ResetCronQueue()         { m.systemInput.CronQueue = nil }
